Backend/pyproject.toml
[tool.poetry]
name = "placeecho-backend"
version = "0.1.0"
description = "PlaceEcho Backend API"
authors = ["Alex <alex@example.com>"]
packages = [{ include = "app" }]

[tool.poetry.dependencies]
python = ">=3.11,<3.13"
fastapi = "^0.128.0"
uvicorn = {extras = ["standard"], version = "^0.40.0"}
pydantic = "^2.12.5"
python-multipart = "^0.0.21"
python-dotenv = "^1.2.1"
httpx = "^0.28.1"
openai = "^2.15.0"
pillow = "^12.1.0"
redis = "^7.1.0"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
